Reference a real attribute, create fake standard file for test payroll models
